Fix bugs generating ORM classes from existing databases
MHK legacy databases don't have foreign key constraints
from the specializing table to the previous level in the
inheritance tree, and this generated a problem with
sqlalchemy. Now the code generates a new column definition
for the id column on the existing specializing table with
the foreign key constraints for sqlalchemy.

Also in the process of generating the ORM mappings now
the code ensures that the super class is mapped before
mapping the specializing class.

This was tested on a legacy database with a few specializations
of 'good' and 'act' (the "santaclara" database)